[#213][reservation] refactor: 클래스명 변경
- factory 에서 Resolver 로 변경
